增加清除 不目录缓存的功能

+++ gdx-psd-tools/src/gdx/keyroy/psd/tools/GdxPsdTools.java
@@ -51,8 +51,11 @@ public class GdxPsdTools {
 		File packFolder = new File(EditorConfig.export_path);
 		send("cleaning", MessageKey.H2);
 		// 清空文件夹目录
-		FileUtil.delete(packFolder);
-		packFolder.mkdirs();
+		if (EditorConfig.clean_folder) {
+			FileUtil.delete(packFolder);
+			packFolder.mkdirs();
+		}
+
 		send("cleaning ok", MessageKey.H2);
 		// 保存数据文件
 		List<PSDData> psdDatas = EditorData.getPsdDatas();
